Fix review comments
1. InsertTableHandle::notNullColumns() is now a `folly::F14FastSet`, so uniqueness is a property of the type rather than a comment plus a runtime check, and TableWriteNode and TableWriter look names up directly instead of each rebuilding a set. The constructor now rejects empty column names. Serialization writes the names sorted, so the serialized plan does not vary with hash order.

2. TableWriteNode::toString() annotates the columns inline instead of appending a separate list after the stats block:

    -- TableWrite[1][test-hive, c0 not null, c1, c2 not null] -> ...

3. PlanBuilder::TableWriterBuilder now takes the connector-specific write request and notNullColumns through separate setters. Passing both used to silently rebuild the caller's InsertTableHandle to graft notNullColumns onto it.

4. TableWriter no longer inspects the input RowVector's own nulls before its children. A top-level input batch never has nulls, so that was a contract to assume rather than enforce.

5. Constructor preconditions now live in PlanNodeBuilderTest, which builds the node with no source data and no connector, and the write-path tests move out of TableWriterTest.cpp into TableWriterNotNullTest.cpp.
