+++ velox/exec/tests/utils/PlanBuilder.h
@@ -545,11 +545,11 @@ class PlanBuilder {
       return *this;
     }
 
-    /// @param insertHandle TableInsertHandle (optional). Other builder
-    /// arguments such as the `connectorId`, `outputDirectoryPath`, `fileFormat`
-    /// and so on will be ignored.
+    /// @param insertHandle Connector-specific write request (optional). Other
+    /// builder arguments describing it, such as `outputDirectoryPath`,
+    /// `fileFormat` and so on, will be ignored.
     TableWriterBuilder& insertHandle(
-        std::shared_ptr<core::InsertTableHandle> insertHandle) {
+        connector::ConnectorInsertTableHandlePtr insertHandle) {
       insertHandle_ = std::move(insertHandle);
       return *this;
     }
@@ -640,9 +640,9 @@ class PlanBuilder {
       return *this;
     }
 
-    /// Unique target columns that must not contain nulls.
+    /// @param notNullColumns Target columns that must not contain nulls.
     TableWriterBuilder& notNullColumns(
-        std::vector<std::string> notNullColumns) {
+        folly::F14FastSet<std::string> notNullColumns) {
       notNullColumns_ = std::move(notNullColumns);
       return *this;
     }
@@ -662,7 +662,7 @@ class PlanBuilder {
     std::string outputDirectoryPath_;
     std::string outputFileName_;
     std::string connectorId_{kHiveDefaultConnectorId};
-    std::shared_ptr<core::InsertTableHandle> insertHandle_;
+    connector::ConnectorInsertTableHandlePtr insertHandle_;
 
     std::vector<std::string> partitionBy_;
     int32_t bucketCount_{0};
@@ -681,7 +681,7 @@ class PlanBuilder {
     bool ensureFiles_{false};
     connector::CommitStrategy commitStrategy_{
         connector::CommitStrategy::kNoCommit};
-    std::vector<std::string> notNullColumns_;
+    folly::F14FastSet<std::string> notNullColumns_;
   };
 
   /// Start a TableWriterBuilder.
@@ -909,10 +909,9 @@ class PlanBuilder {
   /// create a file even if there is no data.
   /// @param commitStrategy The commit strategy to use for the table write
   /// operation, default is kNoCommit.
-  /// @param insertTableHandle Encapsulates information needed to write data
-  /// to a table through a connector. If not specified, tableWrite will build
-  /// a HiveInsertTableHandle with columnHandles, bucketProperty and
-  /// locationHandle.
+  /// @param insertTableHandle Connector-specific write request. If not
+  /// specified, tableWrite will build a HiveInsertTableHandle with
+  /// columnHandles, bucketProperty and locationHandle.
   /// @param storageParameters Physical storage properties of the written
   /// objects, as opposed to the byte layout inside them. Consumed by the file
   /// sink rather than the format writer.
@@ -935,7 +934,7 @@ class PlanBuilder {
       const bool ensureFiles = false,
       const connector::CommitStrategy commitStrategy =
           connector::CommitStrategy::kNoCommit,
-      std::shared_ptr<core::InsertTableHandle> insertTableHandle = nullptr,
+      connector::ConnectorInsertTableHandlePtr insertTableHandle = nullptr,
       const std::unordered_map<std::string, std::string>& storageParameters =
           {});
